Implementation du vrai service pour le login, signup, main et myitems

+++ inquirio-android/app/src/main/java/ca/obrassard/inquirio/transfer/LoginResponse.java
@@ -3,7 +3,7 @@ package ca.obrassard.inquirio.transfer;
 public class LoginResponse {
 
     public boolean result;
-    public long userID;
+    public int userID;
     public String userFullName;
     public String userPhoneNumber;
     public boolean isFirstLogin;
